backend/api-control-asistencia/pyproject.toml
[project]
name = "api-control-asistencia"
version = "1.0.0"
description = "API para el control de asistencia de alumnos, profesores y personal de servicio."
requires-python = ">=3.12"
authors = [
    { name = "Jeiron Espinal" },
    { name = "Mark Timblau" },
    { name = "David Cordoba" }
]
license = "MIT"
readme = "README.md"

dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "sqlmodel",
    "pymysql",
    "python-dotenv",
    "pyjwt[crypto]",
    "passlib[bcrypt]",
    "pwdlib[argon2]",
]


[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-v"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
